Extract methods to generate random data to separate class RandomDataTools
