+++ src/test/java/org/mastodon/mamut/feature/dimensionalityreduction/umap/SimpleUmapDemo.java
@@ -36,13 +36,14 @@ import java.awt.Color;
 import java.awt.Graphics;
 import java.util.Random;
 
+import org.mastodon.mamut.feature.dimensionalityreduction.RandomDataTools;
+
 public class SimpleUmapDemo extends JPanel
 {
-	private static final Random random = new Random( 42 );
 
 	public static void main( final String[] args )
 	{
-		double[][] sampleData = generateSampleData();
+		double[][] sampleData = RandomDataTools.generateSampleData();
 		Umap umap = setUpUmap();
 		double[][] umapResult = umap.fitTransform( sampleData );
 		plot( sampleData, umapResult );
@@ -65,47 +66,8 @@ public class SimpleUmapDemo extends JPanel
 		umap.setNumberComponents( 2 );
 		umap.setMinDist( 0.1f );
 		umap.setNumberNearestNeighbours( 15 );
-		umap.setRandom( random );
+		umap.setRandom( new Random( 42 ) );
 		return umap;
-	}
-
-	static double[][] generateSampleData()
-	{
-		double[][] firstPointCloud = generateRandomPointsInSphere( 100, 100, -10, 20, 50 );
-		double[][] secondPointCloud = generateRandomPointsInSphere( 250, 250, 10, 50, 100 );
-
-		return concatenateArrays( firstPointCloud, secondPointCloud );
-	}
-
-	private static double[][] concatenateArrays( final double[][] firstPointCloud, final double[][] secondPointCloud )
-	{
-		double[][] concatenated = new double[ firstPointCloud.length + secondPointCloud.length ][ 2 ];
-		System.arraycopy( firstPointCloud, 0, concatenated, 0, firstPointCloud.length );
-		System.arraycopy( secondPointCloud, 0, concatenated, firstPointCloud.length, secondPointCloud.length );
-		return concatenated;
-	}
-
-	private static double[][] generateRandomPointsInSphere( double centerX, double centerY, double centerZ, double radius,
-			int numberOfPoints )
-	{
-		double[][] points = new double[ numberOfPoints ][ 3 ];
-
-		for ( int i = 0; i < numberOfPoints; i++ )
-		{
-			double r = radius * Math.cbrt( random.nextDouble() );
-			double theta = 2 * Math.PI * random.nextDouble();
-			double phi = Math.acos( 2 * random.nextDouble() - 1 );
-
-			double x = centerX + r * Math.sin( phi ) * Math.cos( theta );
-			double y = centerY + r * Math.sin( phi ) * Math.sin( theta );
-			double z = centerZ + r * Math.cos( phi );
-
-			points[ i ][ 0 ] = x;
-			points[ i ][ 1 ] = y;
-			points[ i ][ 2 ] = z;
-		}
-
-		return points;
 	}
 
 	private static class PlotPoints extends JPanel
